Move code to end of method
Seems like there is some kind of race condition, so that if we enter a
chat with a friend that is in "pending approval" state - once this
friend accepts us a session is established but the UI is not correctly
updated (or alternatively a session is never created). This seems to
fix that. Zom issue #313.